Fix a major bug that keeps breaking packaged electron app
The problem (revealed by channeling the stdout & stderr to a log file)
was that electron app does not allow modifying the files or directories
within its packaged app. When the app runs and parses the python binary
file `pydist/app`, it encounters `os.makedirs` call that tries to modify
the current resources directory. This throws an error.

Related: https://stackoverflow.com/questions/60881343/electron-problem-creating-file-error-erofs-read-only-file-system